refactor(wordle): Move backend code to a backend/ directory
